refactor: organize package and rename file
